third_party/blink/web_tests/external/wpt/html/canvas/historical.any.worker-expected.txt

This is a testharness.js-based test.
[FAIL] OffscreenCanvasRenderingContext2D.commit method is removed
  assert_equals: expected (undefined) undefined but got (function) function "function commit() { [native code] }"
Harness: the test ran to completion.
